Added "Greedy Query Filter".  This stops queries that are likely to return lots of results or otherwise degrade the performance of the gnutella network.  This is a route filter, not a personal filter; it affects other users, not just yourself.
Currently we stop "[a-Z].???" and "?" (I'm using regexps here.)
We do NOT stop queries for "mp3" "*.mp3", etc.  Why not?  These queries
may actually be useful.  If someone does one query like this to find what's
out there, that may be the end of their Gnutella searching for a while.
So this is actually a good thing.