Changed the TTL handling policy for query requests and ping requests.  Now, we only pass on the request if the ttl (before decrementing) is greater than 1.  However, we send a reply (if possible) regardless of the TTL.
Our old policy was to respond and forward if the TTL was 1 or greater but
to drop completely if TTL was 0.  This means that we forwarded packets that
we knew would be dropped by the other end!  Our new behavior is the same as
Gnutella 0.56.  Gnotella, BTW, totally ignores TTLs of 1 or 0.  (What a
pain!)

Also, I took out the code to maintain ping/query requests/replies.  This
code wasn't done properly, the statistics weren't used, and Ron is redoing
it on his branch anyway.