Fixed bug in last version of UpdateHandler that caused TWO copies of the LimeWire GUI to be displayed when the "check again box" was false or LimeWire couldn't connect to the server.  The problem was that VersionUpdate was using a timer to start up the GUI if establishing the socket took too long, and VersionUpdate failed to stop this timer when the check() method returned.
The Timer created other potential timing problems.  For this reason, I
eliminated it completely.  Instead, VersionUpdate.check() uses the
URLOpener class to provide timeouts.  These timeouts are also used in
the VersionUpdate.update() method.  To make sure that the GUI starts
if the latter method times out, I move the actual call to update()
from UpdateHandler to VersionUpdate.  This also allowed me to make
several methods in VersionUpdate private.

I'm not that confident in this code.  Ideally, I'd like to do more
factoring to make sure that all methods do one thing only, and do it
simply.  But I'm refraining for now.