Changed files: FileDesc, FileManager, QueryReply, QueryRequest, Response, StandardMessageRouter to be able to accomodate the Rich info sending and receiving. The documentation is pretty clear in the files themselves
Modified LimeXMLDocument, LimeXMLProperties, LimeXMLReplyCollection to be a
able to handle the changes made in the programs

Completed RichQueryHandler.

Extended MetaFileManager this class extends the regular FileManager. The
instance() method in the FileManager now returns a MetaFileManager.

The MetaFileManager lets the regular FileManager handle the non-xml stuff
just as before and creates and the data structures for the meta-info
and maintins these datastructures. It also deligates the querytask to the
regular fileManager and the xml querymatcher and unions the results.

The file is pretty well documented.

o search