Added mechanism to 'preview' downloads before they are completed.  This was implemented by adding a launch() method to Downloader, which launches either the incomplete or complete file depending on the download state.  DownloadView then calls this method instead of launching the file directly.
This system works fine as long as your launching program doesn't try to grab
an exclusive lock on the file.  Winamp seems to work fine, but Windows media
player does not.  BearShare at least seems to have the same problem.

As an additional detail, HTTPDownloader COPIES incomplete files to the
library if it could not move them.  This prevents 'can't move to library'
errors if you are previewing the file when the download completes.
The only downside is that it may waste disk space.  Perhaps we should
add a cleanup mechanism.