Added EditingPanel which extends Input panel. This is used to edit existing LimeXMLDocuments
Changed LimeXMLDocument so that attributes of the top level are not stored in the hashmap. They are 
therefore not returned in the nameValue pair list.

Changed InputPanel - took out 2 useless buttons which served no purpose

Changed MetaEditorFrame - merged addMetaData and editMetaData methods. Reduced redundancy

Changed OuterEditPanel - changed content to be of type InputPanel