ID3Editor : This class removes ID3 tags from an XML string and remembers their values. Later when the other method is invoked. It writes out these values to the .mp3 files with the appropriate offsets.
LimeXMLReplyCollection : added another method mp3ToDisk to help write the ID3 data out to disk. Modified toDisk() so
that if the collection is not audio, it is written out to disk in the regular way. But if it is audio, then the 
ID3 tags are removed first(with Id3Editor) and the rest of the tags are written to disk. 

Deletion of meta data for a non-audio collection happens the same way it used to earlier. But for audio meta-data,
the non- ID3 tags are removed, but we do not kill the ID3 data in the .mp3 file.

MetaEditorFrame : calls either toDisk() or mp3ToDisk() depending on the type of collection


Lines beginning with `CVS:' are removed automatically





com/limegroup/gnutella/xml/ID3Editor.java