XMLTableLine: Added the class to effectively display rich responses.
Response: Added method to deteremine if the Response was a toadNode response. 
This is important because ToadNode puts non-xml data between the nulls of responses

SearchView: Brought the score back. TableLine are now created as TableLines of XMLTableLines
depending on the type of response.

ResultPanel: Replaced Meta-Column with Score Column. 
Added a reference to a LimeXMLSchema, made change to constructor
Renders things as before only if the query was a string query, else rendering
is different for rich responses

ResultPanelModel: Replaced rich column with score column. 
Added reference to LimeXMLDocument. And changed constructor. Changed methods
that find names of columns and values to accomodate rich and regular queries.

StandardSearchView: ShowMetadata() uses a XMLTableLine instead of a TableLine.
addResultTab also takes a LimeXMLSchema, which is set to null for odinary queries.

TableLine: Replaced copy constructor, with createClone() method. 
Handles score instead of Rich

LimeXMLDocument: Deals with attributes of root element to have some real values.

MetaSearchFrame: Changed the way it calls addResultTab it now passes a schema.

ID3Reader: Since the audio schema has been changed, we needed to change the way
the XML strings are created while reading the ID3 tags.