SearchView: Creates 1 Normal TableLine per response and again one XMLTableLine per responses per schemaType.
ResultPanel: Added a HashMap: schema->TableLines. Added a menu to the popup menu
to choose view. Re-arragned code and added method to Re-create a new ResultPanelModel
for a view change. 

TableLine : Changed the match method. If the given TableLine is XML returns 3

XMLTableLine : Overrode the XMLTableLine. Checks the type and schema of the other
TableLine. If not XMLTableLine or differnt schema returns 3.

LimeXMLDocumentHelper: Returns a List of LimeXMLDocument[], rather than a 
LimeXMLDocument[]. Each element of the list corresponds to LimeXMLDocuments of
1 schema type in the given aggregrate string.

The current implementation has poor memory performace (too may tableLines) and 
poor CPU performance (has to recreate a TableLine grouper every time the view is 
changed)