LimeXMLproperties: Added a property XMLDisplayDisplayProp as etc/xml/display
LimeXMLUtils : Now has a method to convert the first letter of a string to capitals
and also a method that takes a canonicalized fieldname and returns a proper displayable
field name

ResultPanelMode: checks the schema of this and finds out of it has properties for
display of Fieldnames. If it does it bases the the resultPanel on these properties.