Fixed bug where queries were always broadcast to hosts sharing no files.  The problem was that the host would (correctly) send us a RESET message but no PATCH message.  MessageRouter would then see that the table wasn't patched and force a broadcast.
The solution is to scratch the isPatched() method of QueryRouteTable
altogether.  Now we only broadcast queries to new connections until we
receive the first RESET messages.  The assumption here is that the
PATCH'es will quickly follow the RESET.  (Yes, this differs from the
suggestion of the query routing paper.)  A result of all this is that
the lastReceived field of ManagedConnectionQueryInfo can be null.