LimeXMLReplyCollection: Is a differnt class altogether now. The basic representation is now a HashMap(it was a List). Changed all the consructors. It now validates that the meta-data is indeed correct for each file. If not it changes the meta-data.
Added vriable count and method getCount()

Added variables like outMap, and changedHash which operate like Editor and are used to
write audio data effectively. 

Now there is an inner class that serializes the map of hashes to XML Documents to disk,
so we are not storing the data on disk as pure XML anymore. 

MetaFileManager: At startup it hashes all the files and creates two hashmaps both
hash hashes to File. One is for mp3 files and the other for non-mp3 files.

Now A LimeXMLReplyCollection may be added to SchemaReplyCollectionMapper even if
the ReplyCollection does not have any entries. For this reason the done and getDone()
are removed from the LimeXMLReplyCollection class

SchemaReplyCollectionMapper: Now if we try to add a replyCollection for
a existing entry, it will clobber the older reply collection instead of appending
it to the older collection.

XMLStringUtils: Changed the method that saves metadata for successfully downloaded files
The methods it calls had been changed.

MetaEditorFrame: Calculates the hash of the file as soon as we intend to edit it
and calls all methods in the LimeXMLReplyCollection on the hashes as per the new
signatures