Issue number: 1 Submitted by: rsoule
there was a bug in the connect method when sending the port
if chat is enabled. The code was not taking into account that
the port might be forwarded.  I changed it to first check with
the SettingsManager to see if the port is forwarded, and then
use the appropriate port.