The goal was to return rich audio responses for ordinary queries. The following classes were modified:
Response: added a setMetadata method, so that meta-data may be added for ordinary
responses.

RouterService: Fixed a small bug that would not necessary make anything crash-
but would waste a little time. The wrong constructor for QueryRequest was being
called - the rich Query was originally being set to the same value as the query.

MetaFileManager: Changed the hashMaps hashToFile TO fileTohash, this because
we want to find the hash of a file quickly to be able to find it's 
metadata qucikly. This is optimized for the common case of ordinary searches
being more prevalent in the network.

LimeXMLReplyCollection: The change in MetaFileManager has an impact on the way this
class searches, loads and maintins data. The loading of non-audio meta-data is
slower than it was before. It is not order the number of files, earlier it was
order the number of documents of that schema.

Initializer : Initializes a MetaEnabledMessageRouter instead of a 
StandardMessageRouter...necessary to make allow ordinary queries to get
rich responses. Otherwise the metadata is never picked up from the QHD.