That last commit had an incomplete log message.  What I meant to say is this:
Tries push downloads if normal download fails.  This turns out to be a
non-trivial change.  I had to replace each occurrence of
RemoteFileDesc with RemoteFileDesc2.  RemoteFileDesc's are removed
from 'files' when downloads are spawned, and added back to 'files' or
a new 'busy' array as appropriate.

Also a cosmetic change: replaced all occurences of
'ManagedDownloader.this' with 'this'.