The goal was to read the Gnotella and bearshare meta-data between the nulls and create audio meta-data from it.
QueryReply: grabs the stuff between the nulls and creates responses with 
that data when it data exists between the nulls.

Response: Added a new constructor that takes between the nulls data. and creates
an xml string of the audio schema using that information. 

SearchView: checks if there is metadata in the QHD, and if not tries to 
create LimeXMLDocuments from the responses, and creates a set of 
XMLTableLine from these. 

If there is data in the QHD, that takes precidence over the data between the nulls.