The goal was to create a regular search string for a pure rich query; and to fix a potoenttial bug while reading between the nulls of responses.
Response: Fixed a potential bug. We now ensure that if we do not 
recognize the format of the data between the nulls, we ignore it.

InputPanel: added a method to get the First rich field to be populated. 
The order of checking the fields depends on the schema.

MetaSearchFrame: Now creates a regular search string from the 
rich part of a query, Iff its a pure rich query.

If no fields are populated at all...it discards the query.