Moved the xml directory into the gui/lib directory. Fixed 2 bugs.
LimeXMLProperties: Set the 
	schemas dir to lib/xml/schemas
	data dir to lib/xml/data
	display dir to lib/xml/display
These are realtive to the dir in which we start the program. So we should
start the program in the gui dir.

StandardSearchView: The image.props is now picked up from the
schemas directory. 
Fixed a bug, that causes a rich search with a schema not associated with
any image to use the image of the previous rich search.

MetaSearchFrame: Fixed bug. If no fields were entered, and search was hit
we would get a null pointer exception.