There were two goals. 1) To make the maps in MetaFileManager thread safe 2) To fix a bug - non mp3 files could not have audio metadata(esp ID3 tags)
LimeXMLDocument: The maps in MetaFileManager are now private and thread safe,
but are passed to the constructor of LimeXMLDocument. We need to be extra careful
with them now.

The audio constructor also now checks if there are any xmlstrings left from
the file from disk after iterating thro the mp3 files. IF there are any entries
left, those are added to the reply collection.(issue 2)

If the collection is audio then while we are trying to save to disk we need to
make sure for non mp3 files we do not strip out the Id3 data.(issue 2)

MetaFileManager: The maps are threadsafe. Added locked methods to access them and
also made them private.

XMLStringUtils: Needed to change because LimeXMLReplyCollection was changed

MetaEditorFrame:Needed to change because LimeXMLReplyCollection was changed