There were two gaols: 1) When a user downloads by choosing a normal table line, if there are corresponding LimeXMLDocuments in other XMLTableLines we should save those LimeX MLDocuments when the download is complete
2) Fix a bug, in the case of an audio Collection with a non-mp3 file, the data w
ould not be
saved. The reason was that the incoming Document did not have an identifier set.

LimeXMLReplyCollection: In the method addReplyWithCommit, when are are adding a
Doc to
the collection and commiting, we need to set the identifier of the doc to the va
lue of the file
. The current value of the identifier in the document may be null or wrong.

ResultPanel: 1) Use the Arrays.equals method instead of doing a byte by byte com
pare, when
checking for equality between 2 GUIDs.
2) Added method getCorrespondingDocs(GUID,index). Since the ResultPanel remember
s all the
TableLines that came in for all different views, if we are given a GUID and an i
ndex,
we look up this datastructure, and find all the LimeXMLDocuments that correspond
 to this
response. And return an array of LimeXMLDocuments.

StandardSearchView: 1) Passes RP to MetaDisplayFrame in the constr.
2) Passes RP to a TableLine when calling toRemoteFileDesc on it.

TableLine: toRemoteFileDesc takes a RP, so that it can get an array of LimeXMLDo
cs
which it can pass them to the RFD.

XMLTableLine : changed signature of toRemoteFileDesc to match the above.