LimeXMLReplyCollection: There were two bugs..which would cancel each other out.
Corrected them both. 
1) The schemaURI of the class was never set. This would cause nothing to be removed
from the SchemaReplyCollectionMapper(which is why we were saved)

2) When a collections document count becomes 0, it would remove that collection from 
the SchemaReplyCollectionMapper. This is bad - add a doc, remove it, and add another one..
would have caused and Assert failure if it were not for 1 above. Fixed it.

SchemaReplyCollectionMapper: the method to remove a ReplyCollection is now private...since
nobody accesses it anymore...but it should still be in the API