The goal was to increase the efficiency of the DuplicateFilter, which was taking up about 23% of all CPU used by LimeWire.
DuplicateFilters: Increased efficiency in 2 ways:
1) Count misses instead of matches. This way we are able to stop the moment we 
have 3 misses (we are sure that its not a "Similar" GUID at this point), rather 
than check all 16 bytes and then check for too many matches. Since 90% of traffic 
was not filtered earlier...this technique performs better by a factor of 5.

2) Earlier we would iterate over previously stored GUIDs. This is less
efficient than accessin the underlying (circular) array. Also in the loop
we just stop when we know we have to stop rather than calling hasNext() on the
iterator