First big commit on the column-selection branch. Committed after it was in pretty good shape anyway.
Removed from CVS since they are no longer needed:
XMLTableLine, ViewChooserDialog

Added to CVS for the new architecture:
TableLineFilter, TableLineSelector, TableLineProjector

Modified:
RouterService: The query method no longer requires the schemaURI, since it was causing
problems. As a result we do not have the ability to create a new RichConnectionThread. But it should be easy to add this feature once we are up and runnigna again.

ResultPanel, ResultPanelModel, StandardSearchView, SearchView, TableLine : 
All modified to work with the new architecture. ResultPanel's model is TableLineSelector
and it delegates some functionality to TableLineProjector and the final underlying datamodel is the ResultPanel. Also TableLine now store an array of LimeXMLDocuments.

MetaSearchFrame: Changed so the new signatre of query in RouterService is used