UploadView: Fixed a bug. The removeRow method in (inner class) UploadList used to close the socket when removing a row from uploadView. This was bad, because in HTTP1.1 . The problem was if a download was complete the UploadView would try to take it out, and would also kill the socket in the bargain - this was bad for HTTP1.1
The method is also called when the user kills the download. In this case its a good thing that the socket gets killed. So the solution was to move the code that killed the socket to the killListener.actionListener.

UploadTest: Wrote tests to check if normal uploads are HTTP1.1 compliant. The tests pass.