Change in interface Uploader.
Uploader: Removed methods connect(), getCloseConnection(), setState() and 
setAmountUploaded() from the the interface. The GUI holds on to an interface, 
and does not need to access these methods. So we remove them from the interface. 
However the backend, which uses these methods, can refer to the concrete class
(ala HTTPUploader).

UploadManager: changed some references from Uploader to HTTPUploader as a
consequence of the change in the Uploader interface.