UploadView: Fixed a bug that I introduced in a previous commit on this branch. The code I commited earlier was causing a classCastException.
UploadManager: Fixed the bug that was was causing the hard limit to be effectively
set to n-1 (where n is the hard limit set by user). This was because 
activeUploads was incremented before testing if we could make this connection, and
then it would fail the test. So the fix was to move the code that increments 
activeUploads to doSingleUpload() after testAndInsert, and decrement it at the end 
of doSingleUpload as part of cleanup.