Fixed the code so radio buttons are added for *all* schemas.
MediaType: made private static variables public, since they need to be accessed 
from NamedMediaType. Also added an important note related to the dpendency between 
NamedMediaType and MediaType.

NamedMediaType: 
1) convertMediaType() now returns a List instead of an array.
2) added static method isDefualtType()

StandardSearchView: Iterates over schemas and creates a radio button for for a
schema if its display string is not a defualt type.