Fixes the 99 percent bug
ManagedDownloader: If the amount left is < MiN_SPLIT_SIZE we used to throw
and exception, now, we futher check if the current download speed is too low
and if it is we try to get another uploader and ask it to upload the 
entire remaining part. 

DownloadTester: added test to check this works.