Solved the 99% bug, need to refine it a little more.
BandwidthTrackerImpl:
1. Now keeps around snapshots from the last 10 clicks.
2. When getMeasuredBandwidth is called it returns -1 if less than 3 clicks have
elapsed. Otherwise it returns the average. This is to ensure we are not to 
quick to judge that a connection is too slow.

HTTPDownloader: Now implements BandwidthTracker, so we have finer grained
information about each downloader's speed.

ManagedDownloader:
1. Does not maintain its own BandwidthTrackerImpl. Rather it works by delegating
the functionality to it's HTTPDownloaders and does cumulation. 
2. ReadObject()/writeObject() do not really serialize BandwidthTrackerImpl 
anymore. But we maintain backward compatibility by using (dummy) placeholders.
3.When we are replacing the last (slow) downloader, we wait until we have
reasonable statistics. If we dont do this, two downloaders can start flapping.