Fixed the bug that makes the annotation code fail when the user enters a sting in a ID3 number field, causes the system to break.
LimeXMLReplyCollection: 
1. Added state variables (values to which state my be set while saving metadata).
2. replaceDoc now returns the document it is replacing.
3. mp3ToDisk() now returns an int (for state) rather than a boolean.

IndentingPanel:
1.Created new method - clearField() 
2. ClearAllFields() uses the above method.

ID3Editor: Method write ID3DataToDisk() now returns an int instead of a boolean.
The returned  int is the state of the write

LimeXMLStringUtils: removed method saveMeta(), this code may be used by LimePeerServer. If we need it again see the changes in MetaEditorFrame.saveMeta()

MetaEditorFrame:
1. saveMeta returns boolean which decides whether or not to close dialog
2. If saving results in a bad state, we clear the fields that were offensive,
and show the user a dialog. Also the changes are discarded.