Started connecting to as many RFDs as we have at the time. First step towards using HTTP1.1.
HTTPDownloader:
1. broke the connect(int ) method into 2 parts. a. connectTCP() and b.connectHTTP(). a. establishes a TCP connection and initilizes the byteReader
and b. sends the reads the HTTP headers.

2. changed the constructor, start and stop are now send the connectHTTP(), beacuse
their values are now determined after the HTTPDownloader is created. 

3. connect() calls connectTCP(0). This is a TODO, since its possible some callers
of this method may assume, that HTTP headers have been sent and received.

RemoteFileDescGrouper: Removed all ^M

ManagedDownloader: 
1. Changed the threading model. Now, one thread is spawned per
RFD. The thread tries to establish the connection and receive it part of the file.
The manager thread waits until one of the threads finishes and tries to 
create more threads if the file is not downloaded.

2. In push-downloads case, the acceptor thread must accept the socket. Earlier only one socket would come in at one time. Now each thread may have sent a push, 
so pushSocket, pushLock etc. have been replaced by a more elaborate scheme. 

MiniRemoteFileDesc: one of the classes needed to work with this new method
of passing messages between threads for the push case.

DownloadTester: changed the way the constrctors/methods are called in other 
classes