Refactored the flow control queues in order to allow future enhancements. Now MessageQueue no longer delegates to FixedSizeCollection directly. Instead it is abstract and subclassed by PriorityMessageQueue and SimpleMessageQueue, which delegate to BinaryHeap and Buffer, respectively.  ManagedConnection instantiates these directly, though it only uses the MessageQueue interface.
I also changed how timeouts work.  Messages are now discarded by
MessageQueue when removeNext() is called.  Hence subclasses need not
worry about timestamps, though they might for efficiency reasons.
As a side effect of this, the query reply queue now has timeouts.  A
drawback is that removeNext needs to convey the number of dropped messages.
Instead of returning a (Message, discard) pair, which would force an
extra allocation, MessageQueue now has a kludgy resetDropped() method
which returns the number of dropped messages.  ManagedConnection calls this
when adding AND removing messages, in order to maintain invariants.