MiniRemoteFileDesc:
1. Solved failing push bug. It was failing because the equals method in 
MiniRemoteFileDesc was not taking an object. So it was not really working (it would cause the hashmap to fail). 

2. Added toString() method.

3. TODO: Unit Test.

ManagedDownloader

1. Fixed GUI freezes. It was freezing because it was calling 
ManagedDownloader.getAmountRead() which was synchronized on this.
The worker threads when accessing the common datastructures were also
snychronizing on this. So we changed the locking. All worker threads
access to common data structures are synchornized on a new downloaderLock. 

TODO: Documention about the new locking.
TODO: set so timeout on sockets when connecting and reset so timeouts when downloading.