There was a bug in the socket behaviour. Earlier we would always set the soTimeout for push downloads and never reset them, so if a pushing downloader stalled for 8 secs we would drop the connection.
For normal downloads we never set the soTimeout. So if a melacious client sent us a response and we asked for it, we could be stalled forever.

The correct behaviour is to set the soTimeout for all connections when we are establishing them, and reset them once the download has started.

Took the code out from DownloadManager that would setSoTimeout for push downloads, and moved it to HTTPDownloader.connectTCP. In HTTPDownloader.doDownload() we reset the soTimeout.