Revterted to version 1.54.4.14 (discarded all the lock splitting changes). The is an easier way to do this. We were splitting locks because we wanted to ensure atomicity of assignAndRequest but did not want to hold this' lock while we were doing this, since that freezes the GUI.
Now assignAndRequest is synchronized on stealLock, and within it we acquire this' monitor when we are modifying shared datastructures. This will not freeze the GUI, since we hold this' monitor for a very short time while we are updating the shared datastructures, also atomicity is guaranteed since we are still synchronized.

Futhrer there will be no deadlocks. We acquire this' monitor only after we acquire stealLock's monitor. We never hold this' monitor while we are waiting for stealLock.