Fixed 2 bugs, and cleaned up the code.
1. Solved synchronization bug. If the first worker removed the entire file
from needed and has not yet added it's HttpDownloader to dloaders, and the
manager in the meanwhile is checking if needed.size()==0 and dloaders.size()==0 (both true at this time). This will result in the state being set to COMPLETE and the manager thred terminating. Solved by synchronizing the termination condition checking part of tryAllDownloads3() with stealLock(which is locking assignAndRequest) AND this (as before). The
order is important to prevent deadlocks.
2. Fixed a bad bug. Earlier, if we caught an exception in either assignWhite() or assignGrey() we would call updateNeeded (except for NoSuchElementException). But its not necessary in the case of assignGrey and infact introduces a bug, since if an excpetion is thrown in assignGrey, we add something to needed which is bad, since we never took anything out of needed while stealing a grey potion of the file, and we did not make the victim stop either. So presumably we are now downloading some part of the file twice. 
3. In doDownload moved call to updateNeeded inside sycnched block.
4. Kept the debug statments added to figure this out.
5. Added important comments.
6. Removed some dead code.
7. changed debug method to log to a file if necessary.