Refactored code from IncompleteFileManager to VerifyingFile
DownloadManager: In writeSnapshot there is no need to call updateIncompleteFileManager, 
since the information is now in VerifyingFile and it's always uptodate. We still need to 
write VerifyingFile.readObject and writeObject.

IncompleteFileManager: 1. Does not rememeber intervals for incompleteFiles, rather remembers
VerifyingFile for incompleteFiles.
2. methods getEntry, and getBlockSize do the right thing and maintin backwards compatibility.

ManagedDownloader:gets/adds information about download progress to VerifyingFile rather 
than IncompleteFileManager.
2. Does not create VerifyingFile if it can get one from IncompleteFileManager (for resumes)
3. Intervals are used in IntervalSet in such a way that the high'th byte will be written, 
however HTTPDownloader does not write the high'th byte. So methods assignWhite and updateNeeded
have been changed to accomodate these changes.

VerifyingFile: Its not practical to create a RandomAccessFile before calling the constructor
of VerifyingFile (it can be called from IncompleteFileManager) so the constructor now takes a 
File and creates the RandomAccessFile.
2. Added methods getBlocks(), getFreeBlocks(), getBlockSize()

DownloadTester: changed because the constructor of verifyingFile changed.