Fixed serialization issues.
IncompleteFileManager: 
1.When we serialize we remember File->List<Interval>,
So we added a readObject that convers the List to a verifyingFile which we use throughout the download. Also added a writeObject method that takes a converts 
VerifyingFile into a List of Intervals, and writes it out to disk.
2. Because of 1. we were able to simplify other methods. 

VerifyingFile:
1. added open(...) method. which must be called before writeBlocks()
2. changed constructor
3. small optimization (saves one allocation in the common case)
4. documentation and format changes.

ManagedDownloader:
1. changes in IncompleteFileManager and VerifyingFile
2. getAmountRead() returns 0 if commonOutFile is null

IntervalSet:
added method getAllIntervalsAsList
DownloadTester
changes in VerifyingFile.