1. While converting List<intervals> to a VerifyingFile, we don't add an interval to VF if interval.high < iterval.low
2. When converting a verifyingFile to a List<interval>  we convert the intervals from fully inclusive to half inclusive( because thats what older
limewires used to store and we account for it in the inverse translation)

3. Solved a synchroization bug. While writing out a snapshot, we iterate 
over the list of Intervals. This may be happening at the same time that we 
are adding/removing intervals from it. To ensure that this does not happen,
we need to synchronize this operation on VerifyingFile. 
This adds an new locking invariant: never obtain IncompleteFileManager's lock when holding VerifyingFile's lock