Added feature to verify completed downloads and tested.
ManagedDownloader: In tryAllDownloads2(), before we move to library, we now
hash the downloaded file and compare with the hash of the bucket we downloaded from. If we find a mismatch, we inform user about corruption, and treat it like
a regular corruption. 

RemoteFileDescGrouper: added method getURNForBucket - to help ManagedDownloader

DownloadTester: Added tests to make sure the feature works. and also changed
some alternate locations tests, to work with the new code.

ActivityCallBackStub: Added method to be compatible with the chaged interface