First pass of server side for upload queueing - untested.
UploadManager: Added queueing capability, with necessary datastructures and loops. BUlk of the work is done in insertAndTest. We have added the ability to make sure that queued clients respond within the window of time we expect. 

Uploader: Added new QUEUED state.

HTTPUploader: changed setState to deal with queued condition. 

Added QueuedUploadState: The class in incomplete.