UploadManager: Fixed bug, when an uploader was taken off the queue we used to call setState on it with UPLOADING - which did nothing. Now we call it with CONNECTING which does the right thing.
HTTPDownloader: Fixed bug in supportsQueueing(), we now loop till the end of the headers
which is OK because this method is called only when we have a 503. So consuming all the headers does no damage. Leaving the headers on the socket is bad, because we want to use this stream again. 

After reading all the headers we check if we have all the info we want, if now we throw IOException which is interpreted as TryAgainLater.

QueuedUploaderState: the writeMessagesBody method must do nothing, we dont want to
write anything after the headers.