Completed client side of queueing.
ManagedDownloader: assignAndRequest returns 0,1,2 for notconnected, queued and connected.
If queuedException is caught, assignAndRequest sleeps depending on QX. If the sleep is interrupted, it returns 0 otherwise 1. The call to assignAndRequest in connectAndDownload is in a loop while the download is queued. Disconnecting or connecting will take it out of the loop.