Fixed a bad bug in downloader. If a queued downloader was killed before it started downloading, it would not be interrupted, so it would remain in queue. Eventually when its turn would come, it would be realize it was stopped and return without closing the socket. The uploader in the meanwhile would keep sending data until the sockets buffer was full.
We solves this by calling interrupt in each thread in stop. This will have no effect on worker threads unless they are waiting for push OR waiting Requery OR sleeping while in the upload queue. In each of these cases it is OK to interrupt the thread since we are going to stop anyway.

Note: sleeping threads which are queued have to call stop on their own HTTPDownloaders if they are interrupted.

Also removed useless debug stack traces.