Fixed a bug that was causing NPE when we should have been in File_Not_Found state.
Before we queue, we dont check the status (except BroseHost) but we should be checking more states - like it could be PUSH_FAILED or  FILE_NOT_FOUND in which case we should not be trying to queue or download. Fixed so that we check that the state is connecting before we try to queue and connect.