Fixed bug where LimeWire wasn't validating HOSTFILE data properly.  As a result, it could interpret random HTML as a valid IP address and store it in the gnutella.net file.  This would cause performance problems because of DNS lookups, especially on startup.
The solution is to add a constructor to Endpoint that strictly validates
the IP address.  This constructor is used in BootstrapServerManager when
reading data from GWebCaches and in ExtendedEndpoint when reading data
from disk (in case of corrupted old files).  Updated tests, and moved
unit tests for Endpoint into EndpointTest.