UploaderTest: had started failing because the Uploader returns 206 for range requests so we had to change the asserts.
UploadManager: In acceptUpload we dont have try catch blocks around all the steps because there is was a bug - if any of the steps threw and IOEx, and was in the queue (like in the second or third iteration) then it would not be removed from the uploadQueue. Now we put all the steps in one try catch and in the finally we perform the idempotent operation of checking if the uploader is in the queue and if it is, we remove it and set state to interrupted.