A first pass at MAGNET downloads using MagnetDownloader, a subclass of ManagedDownloader.  Two changes were necessary to allow MagnetDownloader to requery immediately without throttling:
1.  Changed DownloadManager.sendQuery to ignore the global requery
throttle if the query isn't marked with a requery GUID.  The
newRequery method of ManagedDownloader (and all subclasses) was then
changed to take the number of requeries as an argument.
MagnetDownloader.newRequery(n) does not mark the query on the first
call.

2.  ManagedDownloader was changed to call a template method,
nextRequeryTime(numRequeries), to determine when to issue the next
requery.  MagnetDownloader then returns the current time if
numRequeries==0.

I also refactored ResumeDownloader to take advantage of these changes
instead of manually sending a broadcast in the download(..) method of
DownloadManager.  This means that the first resume download query is
not marked as a requery.

A few cleanups to ResponseVerifier were necessary to make it easier to
score incoming results.  The result is a simplified interface.