Fixed bug that would not allow downloader to get alternate locations unless the download was successful. And added a test for it.
ManagedDownloader: Moved the code that added alternate locations into the finally block. Earlier if an exception was thrown (queued, talx etc) the alternate locations would not get added. 

TestUploader: Writes the alternate locations and content range etc, if busy (must do the same for queued also - this is a todo)

DownloadTest: added a new test to make sure alternate locations are used by downloader if prvided by a failing uploader.