Changed HTTPDownloader to use RemoteFileDesc.getUrl() instead of hardcoding '/get/<index>/<name>' syntax.  This allows MAGNET downloads to specify arbitrary URL's for non-Gnutella default servers.  Also note that we will now use '/uri-res/N2R?<hash>' syntax for regular downloads when the server supports it.  (See getUrl().)  This change has already been made on chord-branch, so I expect merge conflicts at some point.
MagnetDownloader now instantiates URLRemoteFileDesc for default locations.
URLRemoteFileDesc subclasses RemoteFileDesc and stores the value to return
in getUrl().  One difficulty is that ManagedDownloader/RemoteFileDescGrouper
copy RemoteFileDesc's passed as arguments in order to augment them with
extra state in RemoteFileDesc2.  To get around this problem, I changed
RemoteFileDesc2 to actually subclass URLRemoteFileDesc.

Updated tests and added tests for URLRemoteFileDesc serialization.  Changed
TestUploader in the process to store the GET requests.