Chose a proper size for CHUNK_SIZE in ManagedDownloader to ensure good chunking, while minimizing performace loss due to extra headers and latency.
ManagedDownloader: Set CHUNK_SIZE to 20K

TestUploader: Earlier when a rate was set, we would send as many bytes as we needed for one second and then sleep for the rest of the second. This does not work for small chunk sizes because as per the rate if we need less than a second, we still sleep for the rest of the second. Now we use a ThrottledpOutputStream - which has finer grained control and the tests work better.

DownloadTest: in one test we defined RATE but did not set the uploader speed - maybe an oversight. Corrected by setting rate.